[refactor] Board 엔티티 리팩터링
