+++ src/main/java/com/example/demo/domain/board/domain/Tag.java
@@ -1,6 +1,6 @@
 package com.example.demo.domain.board.domain;
 
-public enum BoardStatus {
-    FAKE,
-    REAL
+public enum Tag {
+    seminar,
+    notice
 }
